split Object class into its own file
