+++ hypergraphx/viz/__init__.py
@@ -2,3 +2,5 @@ from hypergraphx.viz.draw_communities import draw_communities
 from hypergraphx.viz.draw_hypergraph import draw_hypergraph
 from hypergraphx.viz.draw_projections import draw_bipartite, draw_clique
 from hypergraphx.viz.plot_motifs import plot_motifs
+
+from hypergraphx.viz.Object import Object
